requirements.txt
flask==1.1.1
ipinfo==3.0.0
pymongo[srv]==3.10.0
pyowm==3.1.1
python-dotenv==0.10.3
numpy==1.18.0
scikit-learn==0.22
pandas==0.25.3
gunicorn==20.0.4
Flask-Caching==1.8.0
tensorflow==2.1.2
apscheduler==3.6.3
pytz==2020.1
geopy==2.0.0
timezonefinder==4.4.1